requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
Pillow>=10.0.0
opencv-python-headless>=4.9.0
torch>=2.0.0
transformers>=4.41.0
ultralytics>=8.0.0
numpy>=1.24.0
google-cloud-vision>=3.7.0
